add copyright headers fix tests
