+++ cmd/thriftbreak/main.go
@@ -1,11 +1,31 @@
+// Copyright (c) 2021 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
 package main
 
 import (
+	"errors"
 	"flag"
 	"log"
 	"os"
 
-	"go.uber.org/thriftrw/internal/compare"
 	"go.uber.org/thriftrw/internal/git"
 )
 
@@ -15,30 +35,20 @@ func main() {
 	}
 }
 
-// NoFileError is returned is we fail to provide any thrift files to check
-// and no --repo option.
-type NoFileError struct{}
-
-func (e NoFileError) Error() string {
-	return "must provide an updated Thrift file"
-}
-
 func run(args []string) error {
 	flag := flag.NewFlagSet("thriftbreak", flag.ContinueOnError)
-	toFile := flag.String("to", "", "updated file")
-	fromFile := flag.String("from", "", "original file")
-	gitRepo := flag.String("repo", "", "location of git repository")
+	gitRepo := flag.String("C", "", "location of git repository. Defaults to current directory.")
 	if err := flag.Parse(args); err != nil {
 		return err
 	}
 
-	if *toFile == "" && *gitRepo == "" {
-		return NoFileError{}
-	}
-	// Here we are going to access git repo to find the previous version of a file.
-	if *gitRepo != "" {
-		return git.UseGit(*gitRepo)
+	if *gitRepo == "" {
+		cwd, err := os.Getwd()
+		if err != nil {
+			return errors.New("could not find current directory")
+		}
+		*gitRepo = cwd
 	}
 
-	return compare.Files(*fromFile, *toFile)
+	return git.Compare(*gitRepo)
 }
